pyproject.toml
[project]
name = "otelproto4research"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Aoyang Fang", email = "aoyangfang@gmail.com"},
]
dependencies = [
    "opentelemetry-proto>=1.25.0",
    "pandas>=2.2.2",
    "pytest>=8.2.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "Apache"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
includes = ["src"]